refactor: move download helpers into `network` submodule
