+++ server/helpers/network/__init__.py
@@ -1,0 +1,3 @@
+from server.helpers.network.huggingface_download import (
+    huggingface_download as huggingface_download,
+)
