+++ server/helpers/network/has_internet_access.py
